[android] Cast RouterType to jint in ParsedRoutingData JNI ctor
The ParsedRoutingData constructor is resolved as "(IZ)V", but the scoped enum
RouterType was passed straight into the NewObject varargs. It works only because
the enum's underlying type is int; make the conversion explicit and consistent
with the jboolean cast on the same call.

+++ android/sdk/src/main/cpp/app/organicmaps/sdk/Framework.cpp
@@ -840,7 +840,7 @@ JNIEXPORT jobject Java_app_organicmaps_sdk_Framework_nativeGetParsedRoutingData(
   static jmethodID const routeDataConstructor = jni::GetConstructorID(env, routeDataClazz, "(IZ)V");
 
   auto const & routingData = frm()->GetParsedRoutingData();
-  return env->NewObject(routeDataClazz, routeDataConstructor, routingData.m_type,
+  return env->NewObject(routeDataClazz, routeDataConstructor, static_cast<jint>(routingData.m_type),
                         static_cast<jboolean>(routingData.m_startRouteNavigation));
 }
